Adjust MVC Structure in AgendaViewController.
[Modify] AgendaViewController

[New] SessionsViewController
