+++ Mopcon/Mopcon/Controller/Agenda/ConferenceDetailViewController.swift
